requirements.txt
python-dotenv==1.0.1
aiogram==3.13.1
dependency-injector==4.42.0
fastapi==0.115.2